# swagger.yml
swagger: "2.0"
info:
  title: LLM Doc Embeddings AI Search API
  description: API documentation for AI search
  version: 1.0.0
...
paths:
...
  /documents/{fileName}:
    get:
      summary: Retrieves the content of a document by file name
      description: Get the text content of a document
      tags:
        - documents
      parameters:
        - type: string
          description: The name of the file to retrieve
          required: true
          in: path
          name: fileName
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            properties:
              content:
                type: string
                description: The text content of the file
host: localhost:3000
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...